api/requirements.txt
asgiref==3.8.1
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
drf-nested-routers==0.94.1
python-decouple==3.8
sqlparse==0.5.3
tzdata==2024.2
